Creation of Draft 2
Dummy me forgot to move my progress  into a second draft folder, so now all of files history from draft 1 are erased, and instead updated to this current version. You weren't really missing much though. The homepage was pretty much barren and the auxiliary pages are currently exactly the same as they were when draft 1 was due.

Also, the dummy images were edited to get a better idea of the alignment of the images.